feat(board): pawn captures, pushes, double pushes and en passant

+++ libs/board/src/lib.rs
@@ -14,10 +14,10 @@ use std::{
     mem::variant_count,
 };
 
-pub const BOARD_SIZE: i8 = 8;
-pub const NORTH: i8 = BOARD_SIZE;
+pub const BOARD_SIZE: u8 = 8;
+pub const NORTH: i8 = BOARD_SIZE as i8;
 pub const EAST: i8 = 1;
-pub const SOUTH: i8 = -BOARD_SIZE;
+pub const SOUTH: i8 = -(BOARD_SIZE as i8);
 pub const WEST: i8 = -EAST;
 pub const NORTH_EAST: i8 = NORTH + EAST;
 pub const SOUTH_EAST: i8 = SOUTH + EAST;
